refactor(coordinates): :art: migrate CoordinateManager from static to instantiated singleton
Refactored CoordinateManager to be an instantiated singleton with instance methods, as opposed to static methods used previously. This is part of a larger effort to migrate to a dependency injection pattern. CoordinateManager.deserializeCoordinate was change to take in a Message rather than String, so that its error message may display the Message id rather than its content.